More mistakes discovered: staging commit
After comparing the results some more with Miele (1970), the Q
values were found to be in a different order of magnitude. A lot of
investigation was needed before the mistake was finally found. The
value of Q that goes to the history array corresponds to the values
before the corresponding restorations, whereas the Q values from
Miele refer to the values after the restorations. This makes a big
difference!

In order to fix this, the flow of the code might change a lot, so
a staging commit is in place.