Quadratic interpolation works great with 9-1!
Going deeper into the calcStepGrad, I decided to try quadratic
interpolation. See the code for details.

Testing only with prob9-1, the results are very good, reaching
convergence in half the number of gradsteps as before, and with no
rejections!

Future work is to work better in the stopping criterion and try
with other problems.