Prob 9-2 is something very different indeed...
After the great success with prob9-1, the hopes were up with the
other problems. Following the natural order, prob9-2 was tackled.

It turns out 9-2 is something very different. It just did not
converge with the simple quadratic interpolation that aimed to
minimize (quite naively, in fact) the objective function (J).
In order to properly understand why this was not working, many
other methods were added (which should have been put in different
commits... sorry), and the scheme was modified to allow manual
input of step guesses.

It turns out that, in the current way of the program, after a re-
jection, the calcStepGrad tries a step value equal to 0.9 times the
previously tried step value (In my defense, Miele is totally unspe-
cific about what he tries when gets rejected!). Most times this
works eventually, after some rejections, but sometimes it goes into
a loop of constant rejection from which it cannot escape, because
as the step gets smaller and smaller, the solution changes so lit-
tle that the same situation arises again and again.

I even tried to run 9-1 again, and managed to prevent the program
from converging by some unfortunate step choices. Unfortunately I
could not reproduce that again in order to properly understand
which condition triggers this behaviour of endless rejections.

Running 9-2 again with more plots, I got into a similar situation
as before with 9-1. After many rejections, the step got so small
that the P values AFTER applying the corrections were smaller than
the tolerance, and still all the I values were larger than I(0),
thus leading into an assured rejection (because there would be no
restoration after the grad step!). This was really not supposed to
happen because there was a mathematical guarantee that the grad
lowers the value of I. The only explanation for this would be a
numerical error. Indeed, this happened when the solution was at
Q ˜ 1e-4, that is, the I value is close to the theoretical minimum,
and the solution more sensitive to numerical errors, but still.
I could not reproduce this behavior again. Very elusive...