New ideas in calcStepGrad
(I wad written a huge text and then the commit failed, so I am
sorry if this message is less informative than necessary. This
is VERY frustrating.)

- Implemented Bernardo's ideas in calcStepGrad. Basically there is
a "limit step search" beforehand (via a bisection that can and will
be improved lately). The manual input for step search can be safely
ignored (by pressing 'q' every time), at least for the cases tried.

- Still haven't figured out what triggers the multiple rejections,
but I finally have controlled conditions to trigger it. At the cur-
rent stage of the code, even problem 9-1 goes into endless rejec-
tions, but this can be fixed by alterring limP (line 29 of
grad_sgra.py) to a value of 1e5 * tolP or higher.

- Guidelines for future commits:

-- the stepMngr object should be re-used after a rejection. It is
the very same instance of the same problem, there is no need to
throw away all the results and get them again after the rejection.

-- The post-rejection step search must be alterred. The easy case
is when P<tolP already but I>I0, in this case the step should be
significantly lowered while I>I0, otherwise there will be a rejec-
tion, guaranteed, because there won't be a restoration.