Re-use of the stepMngr object
The re-use of the stepManager was implemented.
However, running prob9-1 with other values for limP, strange results
were found. With limP=1e2*tolP, the program ran without any rejec-
tions, which is strange, yet positive; but with limP=1e1*tolP, the
program went into a cycle of rejections... very strange...