Proposing a mechanism for multiple rejections
Ok, so searching for reasons for multiple rejections, I turned to
prob9-1 with limP set to 1e4 * tolP, which triggers the phenomenon
aforementioned. Observing the behavior of the step search before
the rejections, I believe the problem to be caused by the decision
to do or not the restoration, based on the P value. When the P
value is below but close to tolP in the beginning of the grad
phase, the rejections begin. This seems to happen because some-
times the decrease in I is too small (due to the natural conver-
gence of the problem!), but the increase in P big enough to make
P>tolP, hence causing a restoration, which increases the I value
more than the decrease of the grad, hence rejecting.

The multiple rejections happen when the P is really close to tolP,
because any grad step selected, however small, will still increase
P enough to cause a rejection. This sometimes gets into a point
where the I values do not decrease at all (sometimes by numerical
problems), but P keeps being raised enough to cause a rejection.

Maybe a solution to this is forcing an “unnecessary” restoration
after the first rejection (or, more adequadely, after the rejec-
tion cycle is finished by a gradAccept). Trying this yielded bet-
ter results, but not so good as to avoid the multiple rejections.
Another possibilty is to maybe even prohibiting a grad when P is
too close to tolP, but this would mean breaking with Miele’s me-
thods more agressively, so it was decided not to do it now.

By the way, comparing two curves via loglog plot is apparently not
reliable when the difference between them is small. More than once
the plot showed one curve being much below the other, but then the
numbers told a different story... Changing the I and Obj plot to
semilogx made things much better! This led to many, many changes in
the plotting of the results, which in turn led to many good insights.

It was noticed that the calculations of J and I were employing dif-
ferent integration schemes (which is just plain wrong, and was cor-
rected immediately). Then, I got a glimpse of the real problem: the
lambda integral which makes J different from I is also wrong. The
error term, dx/dt-phi, is calculated in the middle of the steps, and
not on the steps themselves. This produces a relatively big source for
numerical error, in the calculation of J (which is, of course, more
significant when the I/J variations are naturally small due to the
convergence process), and I do believe that is the source of this
erratic behavior.