It was a sign after all...
So, after thinking a little about how to adequately calculate the inte-
gral L in J, I noticed a mistaken sign in the definition of L, (dx/dt-
phi) vs (phi-dx/dt). That turned out to make prob9-1 work perfectly, wi-
thout any rejections, even with limP set to 1e2 * tolP. With 1e1 * tolP
there were problems, but hey, it is a limit way too low after all.

All the other problems were tried: 9-2, 10-1, 10-2, probCart, probBrac,
probZer and it worked for all of them. Some values of the limP parame-
ter were also tried, and it turns out the program is rather unsensiti-
ve to it. It does, however, show the expected behavior of becoming more
“cautious” with a smaller limit, and hence taking more grad iterations
to converge. Not that many more, though.

I do believe we are ready to tackle probRock again!

In the next commit, I will put the limP parameter in the external con-
figuration file (.its) for each problem.