Staging before new idea
As the parameters were being included in the configuration files, I
noticed that the calcStepGrad is not too good in the automatic mode,
more specifically when the local minimum in Obj is not found and the
program should return the maximum value of alfa so that no restric-
tions are violated. By the way, with the parameters used, this only
happened in prob10-2, mind you.

Therefore, I am about to change things a little bit, so a commit is
in place.