Reaching for the two milestones before probRock
Ok, so thinking better about the situation, it would be better to
solve two more problems (milestones) before switching to probRock:

- the limitation (used in probRock for the acceleration limit)
still has not been tested after the reform: always best to test
these new functionalities in simpler problems

- the solution saving/reloading functionality has to be readapted
due to the new flow of the program: this is specially useful in
probRock to avoid having to run itsme at every single time sgra is
run for tests and stuff.

So, with this in mind, I proceeded to probLQR, as the most tunable
problem already implemented. After finding out an error in the I
integration (generalized automated tests would be great, by the way),
I ran probLQR with many different settings and discovered that there
is not a unique solution for the LQR as it currently is. It also
turns out that the current formulation of the LQR problem did not
match any version of solved LQR problem online or in Kirk.

Hence, the next commit should make the LQR problem even more tunable.