probRock just finished running
Yep. That's it.

While searching for the second milestone, I thought it would be
better to try it with probRock itself, since other problems seem
to run pretty fast. After adequating probRock to the new confi-
guration variables, I left the program running overnight. And it
worked. Sure, the acceleration limitation was in a very safe va-
lue of 35g, but hey.