MAJOR fault in rest discovered!
After changing a lot of simple things to increase the performance of
probRock, as well as removing some minor plotting bugs, I bit my
tongue and found a major fault in rest. For the first time, the rest
rounds were somehow increasing the value of P!

This calls for a rest-reform. Let's try to keep it simple though.