Pi limitation in cart and rock, but no test
Running probRock some more times, increasing GSS_PlimCte, it was
noticed that some arcs became negative (pi<0.0). I recorded that
in the original paper (1970), Miele did say the grad step search
should make sure pi>0.0 when pi is the time of the arc. So, this
extra condition was included in the the grad step search, very
generally, in fact. The user may input the desired limits for the
pi values as desired.

I could not properly test it, however. Even setting GSS_PlimCte to
an absurd value, probCart would never compute a solution with
pi<0.0 in either arc. Maybe other problems would...

With probRock, setting the minimum pi condition to 0.0 did not
work, there was a zero-division error when calc step grad tried a
step below 1e-300. Some kind of saturation would be important here!

But first, the other problems' default files must be corrected.