Pi lims in other problems
The other problems have been modified to allow pi limitation. In
most cases, only the .its file had to be changed, but LQR had to
be changed a little bit more because of a minor bug in the parameter
loading section. I used probLQR to test a little bit more the pi li-
mitation. It seems to work fine, but there is a situation here.

If a given sol is already violating a pi limit (maybe because of a
bad initial guess), then the calcGradStep cannot go anywhere; even
alfa=0.0 gets rejected as a bad step, and the whole algorithm
breaks down. Sure, this bad initial guess issue can be “solved” ra-
ther trivially by checking the pi condition beforehand and refusing
to go further in case of violation. But if the restoration gets a
non-violating solution and turns it into a violating one, we are in
serious trouble.