Integrations mismatch and negative times
I noticed the f function integration for calculating I was still done
with the trapezoidal method. When I changed the code for using Simp-
son, the program failed to converge. The difference between yielded
values for the integration methods is always smaller than 0.1%. This
program is way too unstable…

Besides, the first arc is getting a negative time in the last steps…
This probably happens during restoration, because the grad is (or at
least should be) explicitly forbidden to let any pi become negative!

There is still a LOT of work to be done.