Major integration mistake in calcP
While looking for a way to prevent a piLim violation during rest, I
noticed a potentially huge mistake in calcP. It was due to switching
from the trapezoidal integration to Simpson integration in the calcu-
lation of P_int. The mistake was scaling down the value of P_int by
1/(N-1), which makes rest think P is actually much smaller than it is.
After fixing the mistake, I tested all the problems again, and the
performance has not changed at all.

ProbRock, however, shows another behavior entirely. The program failed
to converge completely after removing the aforementioned P_int calcu-
lation mistake, even as the ‘I’ calculation was reverted to the trape-
zoidal method. It only ‘works’ with the P_int mistake AND the trape-
zoidal ‘I’ calculation together. This suggests that the convergence is
simply due to coincidence. This is corroborated by Muniz’s claims that
the program fails to converge with different rocket parameters such as
Isp.

On the other side, it was also discovered that the pi limitations were
not active at all, since the limitations were set to ‘None’ in the con-
figuration file. I must now include these changes and test everything
with probRock again...