Back to square one...
Setting the lower limits to 0.0 did not work at all. The program as is
runs very normally, optimizing the solution, up to a very specific
point (I ˜ 1512.4, Q ˜ 0.11, P.M.G. ˜ 29.1%), where the gradStep gets
very small (˜1e-12), which triggers rejections due to unsuficient lo-
wering of ‘I’, and the known dead end.

This very specific point where the program stops optimizing the solu-
tion is independent of the main variable that controls the optimizing
rate, GSS_PLimCte. I tried running the program with GSS_PLimCte =
[1e3, 1e5, 1e7, 1e9], and in all of the runs the program stopped at
the same “point” during the convergence (at different iterations, of
course):

Table 1. “Small steps” transition point vs. GSS_PLimCte
_______________________________________________________
GSS_PLimCte   Iter      I       Q    Payl. Mass Gain
    1e3      ˜1100   1512.4   .117       29.07%
    1e5      ˜320    1512.4   .117       29.08%
    1e7      ˜100    1512.4   .117       29.14%
    1e9       36     1512.3   .102       29.15%
———————————————————————————————————————————————————————

Running the program for a rocket with 30kN thrust generated a similar
problem, but the program stopped in a different point (as it should,
since the initial value of ‘I’ was already lower than 1512.4). With
a single rocket stage, the same result appears, and turning back to
trapezoidal integration for ‘I’ and the vetIP *= dt mistake, the pro-
gram also runs! Muniz even tried the single stage case with only the
trapezoidal integration for ‘I’ and it worked.

This suggests that the trapezoidal integration in f works better with
the rocket problem in general. Indeed, in lmpbvp, the integration me-
thod for phiLamIntCol is the trapezoidal method… I will try this in
the next commit.