Another huge integration mistake
Ok, so after the P_int integration mistake, another mistake, similar
and equally shameful, was found. It turns out all the Q components
were being integrated in the same wrong manner.

After fixing it, I decided to review all the integrations on the pro-
gram, converting all of them to Simpson. After that all the problems
were tested again, and the convergence rate was lowered in all of
them. This should not come as a surprise, after all, due to the inte-
gration mistake, the Q value was being lowered to a factor of 1/(N-1).
Problems 9-1, 9-2, 10-1, 10-2 and cart had very small increase in the
number of grad iterations (no greater than 5); problems brac and LQR
behaved very differently. Indeed, in these last two, N was set to 500,
so the integration mistake was effectively loosening tolQ by 3 orders
of magnitude.

An interesting effect was found in probBrac: rejections followed by
convergence. Using the debug option “plotCalcStepGrad”, it was possi-
ble to find out that the optimal corrections (in the sense of minimi-
zing J) were getting smaller and smaller, until the point where the
lowering of ‘I’ was very small but the ‘P’ values were still raised
beyond tolP, so after restoration, the ‘I < I_old’ condition did not
hold. For this problem, lowering tolP from 1e-8 to 1e-12 avoided the
rejections completely. Another interesting result was obtained by in-
creasing N from 501 to 5001: it took 3 MORE grad iterations to con-
verge in the large N case. That is rather surprising, and suggests
that increasing N does not really make the convergence faster, in
terms of number of iterations.