New experiments, new insights
This whole commit is dedicated to studying a strange phenomenon, pre-
viously reported, here referred to as “the wall”: The program works
normally, lowering Q, and then, with or without rejections, the step
values suddenly get very small (<1e-10), and the program stalls (i.e.,
“hits a wall”.)

The first variable that influences this effect is GSS_PLimCte. Firstly
the value used was 1e9; with this value, the program hits the wall a-
round GradIter #32. Plotting the I and J histories during GSS, it can
be seen that the J values indeed are not lowered with usual values of
step (˜1e-4), but they would be if the P>limP condition was not active.

So, the program was run again with an absurd value of GSS_PLimCte,
~1e20. In the first GradStep the solution entered in an infinite res-
toration loop, probably because the rocket entered the Earth, or the
speed became negative, or something like that. Indeed, the P limita-
tion is really necessary.

Running the program again with GSS_PLimCte = 1e11, the wall is hit
earlier, around GradIter #13. Looking at the I and J histories during
GSS, weird behavior was detected. Most of the times, the chosen step
did not correspond to a natural local minimum of J, due to the P limi-
tation. Sometimes the chosen step did correspond to a natural local
minimum of J, but this was not a local minimum at all for ‘I’, in fact
sometimes this chosen point was such that I>I0. Even more strangely,
after restoration, the I<I0 condition was met. Maybe minimizing J is
indeed very wise...

When the ‘wall’ was reached, though, the rejections began. Plotting
also the “theoretical” or “expected” J variation (by using the
dJ/dStep derivative at zero, calculated analitically) shows the “non-
linearity” of the problem. The weirdest thing is that sometimes the J
and the ‘I’ variations are POSITIVE even for relatively small steps,
like 1e-5 or 1e-6. On the other hand, the correction is so that the
mass variation is positive in the end of the last arc; therefore ANY
positive step value should produce a negative variation on ‘I’ (at
least before the restoration), which was NOT observed. This unconsis-
tency can only be explained by HEAVY non-linearity in the regions
where the thrust alterations are present.

This situation is not completely avoidable, but it can be diminished
if the controls are desaturated in the initial guess. After all, the
problem is mostly non-linear near the control saturations. Some pre-
liminary experiments in that sense have yielded good (although not
exactly great) results.