Arc separation at target height, plot upgrades
After desaturating the initial guess, the same problem appears (later).
The arc separation rules were changed to keep the 2 initial arcs at h =
0.1 km and h = 2.0 km, respectively. This results in adding two more
psi components. For now, this was done in a terribly hardcoded fashion,
but at least the program was shown to run under these conditions. Now
comes the (boring) part of generalizing the psi and psi_x calculations
on the code.

There were also some plotting upgrades such as including an adimensio-
nal time version of comp.pdf and decreasing the size of the stage mar-
kers in the plotCat method.