More plot upgrades, crazy tests... sucess?
[DISCLAIMER: There is a (stupid) bug that broke the compatibility
with all the problems other than probRock. I am only committing this
now because Muniz asked me to. Issue #84 was created to remind me to
fix this.]

Still investigating the “wall” problem, I decided to include the
accumulated plots of the integrands of Qu (and Qx) to gauge better
the impacts of the “residual peaks” in the Qu parcel. It turns out
the impact is rather small due to the narrowing of the peaks. We
still believe that it is these peaks that are causing the wall
problem, though.

When discussing the problem with Muniz, it occurred to me that this
could be a grid issue, that is, that the time array was not suffi-
ciently fine in that region. I tested this hypothesis in the easiest
way possible, increasing N from 501 to 5001, with some interesting
findings. First of all, the P value after the desaturation was much
higher, but was still brought down with the same ammount of rest it-
erations. The wall problem still occurred, although later (around
Grad Iter #20). Observing the “peak formation”, it was noticed that
the peaks get shorter and thinner, up to some point, and then they
stop. With N = 5001, the peaks increase again, remaining thin, be-
fore the “wall”.

It is concluded that, fortunately, increasing N does not solve the
“peak generation problem”. It was decided then to plot the residuals
with the correction applied, just to see what happens. It turns out
that the Qu term does not necessarily decrease right after applying
the correction, so this does not really make any sense.

Two other changes were introduced: an “oblivius” desaturation, that
worked but still has to be improved, and a somewhat dishonest appro-
ach for calculating ‘I’. Since the integral was not accurate enough,
and since the grid could not solve the problem, I simply replaced
the integral in ‘Iorig’ for the mass difference penalized by the
structural efficiency. It is not as fare as changing the entire for-
mulation, replacing f for g (final cost term), but it seems to have
worked! The Q values were lowered to a record of 2.04e-2, and the
program is still running after 1312 GradIts with NO REJECTIONS.
Still, we all know better than to proclaim victory this early...