Extra prints, but no Miele validation
I've put on some extra prints, such as the Miele-comparing initial
parameters, and the mass distribution table in plotSol, but still
cannot make itsme generate an initial guess to a case closer to
Miele's monostage rocket. The thrust-to-weight ratio of 1.3 is just
too small, apparently.

There will be some issues as well when dealing with acceleration
limitation. I will now try to work on the generalizing the psi con-
ditions in order to compare with multistage cases.