No more hardcoded mess in psi and psiy
Ok, so the hardcoded mess in psi and psiy was generalized for the
multistage case, allowing unlimited natural arcs and "fixed height"
arcs. The only constraint is that the "fixed height" arcs must come
before the natural ones.

Running probRock, I was rather surprised to see it not working in
the multistage case. It does not seem to be "hitting the wall",
but the step values get lower and lower (continuously) until the
program does not go any further. I have to address this problem.

On the other hand, itsme is not behaving very well with the low
thrust to weight ratios required for running the exact Miele cases.
I have tried, without sucess, to understand the itsme code to try
some improvements on my own; and now asked Souza to look into it.