Experiments with low TWRs
Still trying to acheive the thrust to weight ratio of 1.3 in order
to validate Miele's results, I followed Souza's recommendations to
remove (comment) the line that aborts itsme in case of TWR<1.0.
This indeed improved the capabilities for lower TWRs, (sometimes
the initial trajectory was off by a few hundred km, but rest mana-
ged to "bring it back") but itsme still could not acheive the tar-
get level of 1.3. This issue must be fixed! Please, Souza, you're
our only hope!

Meanwhile, there are some sgra-related questions to be answered,
so I will address the other issues.