Added history for GSS stop motive (issue #80)
Fixed issue #80. The ticks for the graph are not great, but this
works for now.

I tested this by running probRock classic monostage problem. It is
interesting to notice that there is ONE single grad step where the
GSS stop case is due to excessive evaluations, even when the maxi-
mum number of obj evalss is raised to a huge value (e.g., 5000).
This was simply not understood.