New problem (lander)
I decided to add this new problem, that tries to perform soft
landings of spacecraft with throttable rocket engines. It turns
out that the optimal solution for this problem is a fully satura-
ted "bang-bang" rocket version of probCart's optimal solution, in
which the spacecraft is initially left at free fall for some time,
at which the engines are turned on at full power.

The problem seems to be very sensitive as well. I had to build a
function for setting a good initial guess, so this commit is actu-
ally a little late.