Now probLand is working!
After removing all the bugs, probLand is finally working.

The initial guess function assumes two phases: one for free fall,
followed by a powered descent final phase. The program worked for
two values of TWR for the final phase (for the initial guess).

Next, I will adapt the program for two arcs, one for the free fall,
in which the rocket will be forced to keep the thrust at zero, and
the other "free" one.