Two arcs for probLand, further generalization needed
Now probLand has been modified for two arcs, the first for free-
fall and the other for powered phase. It did not serve the purpose
of testing sgra as intended, though. I need to generalize it even
further so that it accepts 3 or more arcs, for a truly suboptimal
initial guess.