Arbitrary number of arcs for probLand
Ok, so finally probLand is OK for an arbitrary number of arcs.
I tried running it with four arcs, representing two sections of
free fall intertwined with two propulsive, stopping sections.

It turns out there is also a wall in this kind of problem, with the
step size becoming smaller and smaller even though the Q residuals
are not zero. Let's dive in!