Tiny generalization, big results
I performed a tiny generalization, changing the free-fall arcs for
thrusting arcs, but with 0 <= thrust to weight ratio < 1. It turns
out setting the TWR to a small value, i.e., 0.1, managed to revert
the "vanishing arcs" problem, and the optimal solution was found.

However, setting the TWR to 0.9 made the other arcs (TWR>1.) start
to vanish! I do think that these "loose arcs", that is, arcs with
no physical constraint, like h(1,i) = 2km = h(0,i+1), are somewhat
dangerous. First of all, they do make the problem become mathema-
tically ill-posed, with no unique solution. Numerically, that can-
not be good.

Next, I will try to include an option for physical separation of
the arcs, similarly as the original rocket problem.