Physical separation makes probLand converges wrong... like probRock!
Ok, so I implemented the physical separation, separating the arcs at
h = ˜6.35, 5. and ˜1.35 km, respectively. This worked very well for
avoiding the arc collapse, as expected, but produced an even stran-
ger result.

The initial guess was apparently too weird for sgra to solve. It did
converge, finishing the execution with Q = 6.7e-6 and I = 120.84;
but the final solution was not the optimal one. The previously opti-
mized solution with two arcs yielded I = 117.15 with Q = 8.0e-5.
If that was not unsettling enough, both the histQ and histGradStep
are very much alike to the ones where probRock converged, rapidly
oscilating in gradStep and Q near the end. Yes, this suggests that
the converged cases in probRock are not truly converged. Great.