Sine parametrization for thrust = disaster for probLand
Due to the poor performance of the program in probLand with a "bad"
initial guess, I tried changing the control (thrust) parametriza-
tion, from the hyperbolic tangent:

T = T_max/2 * (1 + tanh(u))

to the sine:

T = T_max/2 * (1 + sin(u))

and the results were pretty bad. Indeed, the program did not con-
verge at all for both the case with the "bad" initial guess and
the case with a good initial guess, in which it converged with the
tanh function. It was actually rather hard to make the code more
flexible in order to allow different control parametrizations, and
due to the poor results, I don't think I will be doing this for
probRock as well.