Fixed tiny bug for probLand, more cases converged
There was a small bug (a programming mistake, actually) concerning
the saturation in calcAdimCtrl(), which was fixed.

I also ran probLand with other values of psiHigh and psiLow, and
they converged. It is clear that the current formulation of MSGRA
is way too dependent of the initial guess, specially if there is
some kind of saturation involved. It seems we need indeed to keep
working in a different initial guess, one that gives a very weird
trajectory, but with controls far from saturating (and probably
respects the acceleration constraints as well).