Odd realization with good old probRock
Muniz showed me another Miele paper, this one from ’98. It
sheds some more light into some specific problems involving
probRock. Namely, it explicitly avoids the V=0 singularity in
the beginning of the launch by forcing beta = 0 and a straight
trajectory. Since the 2003 paper points to this one for details
in the aed model, I bet some other ideas came from there as
well. Anyway, the aed model involves CL and CD varying with
Mach, so if we are going to validate the code with Miele (’03),
we will have to implement that too.

Anyway, with or without validation in mind, the previous re-
sults with probLand make it clear that a fully desaturated ini-
tal guess is required for achieving a truly optimal solution.

So, this branch was created to make a new naive initialization
for probRock. First of all, I needed to gauge the robustness of
rest, in order to have a feeling of how messed up the initial
guess given by the naive method could be.

Unfortunately, though, running probRock again without any
forced desaturation gives a solution with great values of both
P_int and P_psi due to low smoothness of the thrust function in
the current version of itsme, as well as poor numerical integra-
tion. Rest still manages to bring the P values down, but to do
so it messes up the solution pretty bad, putting the angles of
attack to large saturating values for a long time. This gives
grad a pretty rough “initial” guess to work on, leading to the
known behavior. The program does still “converge”, though,
with the familiar Q oscillation that I mentioned in Nov 29
(commit ef765c4e).

So, yes, this is the state of things now. Muniz and myself will
now begin working in a new (less) naive method. I also tested
removing the gamma factor that prevents turning during the
first (traditionally, 3) seconds of flight. The program also con-
verged, but with much more iterations. We still have to decide
how to approach that (time based / speed based / height
based fixation), so it will be nice to leave the possibilities
open in the code.