Naive failed the rest test
With some very small changes, the naivRock now generates a solution
and is able to try and restore it afterwards. The current solution
failed the restoration test; and actually revealed a bug in
calcStepRest, leading to issue #89.

I am already thinking in an enhanced version of the control for the
last arc, hope I can finish it soon.