Naive worked! Not robust at all, though
It worked! At the end, simple is better than complex, I presume.
The controller for the final arc is simply a kind of "deadbeat"
controller that aims at nullifying the dynamics and making both
speed and flight path angle converge linearly to the orbit values.
The time for the maneuver is calculated integrating the height
equation with the v and gamma profiles and making it equal the ne-
cessary increase in height.

This was the first time that naive yielded an initial solution that
could be properly restored (just 4 restorations were necessary).
However, this method is not at all robust, for when I changed the
rocket initial mass from 3000 to 4000 kg, hoping it would not make
much difference because in both the second and third arcs the thrust
is calculated by TWR, but instead the program blew up completely.
There should probably be an iteration loop for the second arc, for
adjusting both the TWR and the constant turning angle.