naive adapted for MSGRA, results OK but wall was hit
The naive module was adapted finally for use in the main program,
and some changes had to be done as well in interf.py and
probRock.py. The program worked OK, producing a coasting section
out of nowhere and bending the trajectory "down" as it did before.

Unfortunately, the wall was hit again. I used to think this was
due to the thrust saturation in the initial guess, but it seems
that is not the case. This is a problem that must be solved!

Happy new year!