Switching back to reform branch for updates in calcSteps
After running msgra many times again, the wall hit problem still
eludes me. It may be a numerical problem, since the formulation
assures both J and I should go down with sufficiently low steps.
Tested the same case with several values of N produces different
behaviors every time, but does not quite converge.

I wanted to keep checking the dI/dAlpha and dJ/dAlpha conditions,
but currently calcGradStep "corrects" the values of J, making this
check almost impossible. Hence, I will now go back to the reform
branch in order to work in the calcSteps.

Another thing that must be done is to change naivGues so that it
stops saturating the angle near the end. But this will have to
come later.

I also added a module for post processing. The idea is to be able
to load a solution (converged or not) and use its plotting methods
in interactive sessions, or produce new plots such as the I vs. Q
plot. This will come in handy when we start discussing proper
tolerances for Q in the future.