Added mass calculation for naive + lots of experiments
The only major change to the code is a function for calculating the
initial mass of the rocket in naive. It is not even being used right
now.

The first experiment was to try to understand the bizarre results
reported in commit ef765c4. Everything points to a grid-related issue,
associated with large saturating regions, where the Qu residual tends
artificially (but expectedly) to zero.

The second was to change, rather dramatically, the search algorithm in
calcStepGrad in order to try something similar to the current one in
calcStepRest, limiting the gradStep to 1. It was a disaster, failing
to converge in prob9-1 and taking >700 gradIts in probCart.

We also tried messing with the tolerances and GSS_PLimCte, to see if
it would make any difference. Apparently, the wall was hit in all of
these cases, or at least kind of. In this last version, the gradient
step does not suddenly get small, nor there are suddenly many
rejections; the step values curve just becomes "noisy". However, aside
from some small number of rejections, the GSS algorithm seems to be
finding (with local minimum) smaller and smaller steps (with the
mentioned noise) by minimization.

I still haven't modified the calcStepGrad in order to actively compare
the I and J variations to the "theoretical" first order ones, and I
still have one idea left for changing the algorithm, but I have to
merge Souza's last modifications to itsme.