Included I and J "relative reduction" calculation, more insights
After a long delay, I finally included the J "relative reduction"
calculation, that is, the reduction J(step)-J0 relative to the
theoretical reduction -Q*step: J(step)-J0/(-Q*step); as well as
its analogous value for I. The interpretation is as follows:

- If this value is close to 100%, then the first order terms are
more significant than the others (as expected) and the reduction on
the cost is going as it should.
- If this value is higher than 100%, then the higher order terms
are helping the cost reduction, which is rather lucky but OK;
- If this value is much lower than 100%, then the higher order
terms on I or J are preventing the cost reduction. This is what
happens when the solution stops converging, by the way.

This brings very interesting insights for analyzing the code,
specially because the relative reduction is strictly decreasing for
the naive initial guess, but follows a more "erratic" pattern for
the itsme initial guess. Of course, all these comparisons are
somewhat inadequate because naivRock does not get its parameters
from the .its file.

Also, rather silently, the constants alpha min, alpha max, beta min,
beta max, etc were turned into arrays. The idea is to allow
different values for each arc, specially the first one, so that the
"gamma damping constant" can be removed later.