Added Hamiltonian calculation and minimization check
(Actually, I had already added it, due to a staging mistake...)

Many problems in SGRA use control limitation, which is handled in
Miele's original formulation with a limited mapping (such as sine,
or, in our case, hyperbolic tangent). This is known to induce
"false positives", in the sense of nullifying the Qu residual in
the sections where the control is saturated (in any range).

Hence, in order to try to make a more robust optimality test, I
added these methods for calculating the Hamiltonian and checking
effects of variations on the control. Interestingly enough, all
the problems with control limitation (brac, cart, rock) failed the
optimality test after MSGRA had decreted convergence; showing the
innate fault of this formulation. Of course, none of this matters
if the changes to obtain TRUE optimal solutions are so small that
the optimal cost function is altered, say, on the 4th or 5th digit.
But it sure is hard to discover when it is the case.