Added caseGen.py: .its generating code
After discussing possible alterations to itsme, Souza advised me to
adjust some parameters in the .its so that the inferred parameters
would be the same both in itsme and naive initial guesses. However,
since I am utterly terrible at doing these adjustments by hand, I
wrote a small code to do it for me.

For now, the only thing that is "working" is the payload ajustment
to yield a given value of initial rocket mass as specified. For some
reason, itsme is very sensitive for monostage rockets... It just can't
seem to converge when the structural efficiency is set to 90%, even if
the payload is very small.